requirements.txt
streamlit==0.52.2
numpy==1.18.2
Pillow==6.2.1
keras==2.3.1
tensorflow-cpu==2.6.0